Commit con votos

+++ src/main/java/com/example/demo/gestionEquipos/infrastructure/entity/JugadorEntity.java
@@ -20,7 +20,7 @@ public class JugadorEntity {
 	private String nombre;
 	private String apellidos;
 	private int dorsal;
-	private float valoracion;
+	private double valoracion;
 	private int goles;
 	private int asistencias;
 	private String equipoId;
